pyproject.toml
[project]
name = "sp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=5.5.2",
    "cmdstanpy>=1.2.5",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "lxml>=5.3.2",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "polars>=1.26.0",
    "prophet>=1.1.6",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "torch>=2.6.0",
    "transformers>=4.51.0",
    "uvicorn>=0.34.0",
    "yfinance>=0.2.55",
    "tavily-python>=0.3.0",
    "numpy>=1.26.0",
]
